Mais ajustes no projeto.
- Tabela de descontos criada.
- Ajustes nas rotas.
- Ajustes nos controllers.
- Ajustes nas migrations.
- Ajustes nos models.
- Tabela de usuarios renomeada para clientes.
- Ajustes nas validações dos forms.
- Feita a exclusão de arquivos desnecessários.
- Feita a função para aplicar desconto no carrinho de compras.

+++ database/migrations/2024_12_11_163030_create_products_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('name')->nullable(false);
             $table->integer('stock')->default(0);
-            $table->float('price')->default(0);
+            $table->decimal('price', 50, 2)->default(0);
             $table->timestamps();
         });
     }
